% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoolAir.R
\name{CoolAir}
\alias{CoolAir}
\title{Function to get relative humidity/condense note after cooling}
\usage{
CoolAir(t.pre, rh.pre, t.post, temp.list)
}
\arguments{
\item{t.pre}{Original Dry bulb temperature}

\item{rh.pre}{Original relative humidity}

\item{t.post}{Final Dry bulb temperature}

\item{temp.list}{the stored data from psychrometric chart}
}
\description{
Function to get relative humidity/condense note after cooling
}
\examples{
CoolAir(78, 45, 56, data.chart)
$rh
[1] 97
$note
[1] ""
}

