% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooling_air.R
\name{cooling_air}
\alias{cooling_air}
\title{Function to get relative humidity/condense note after cooling}
\usage{
cooling_air(T_pre, RH_pre, T_post, temp_list)
}
\arguments{
\item{T_pre}{Original Dry bulb temperature}

\item{RH_pre}{Original relative humidity}

\item{T_post}{Final Dry bulb temperature}

\item{temp_list}{The stored data from psychrometric chart}
}
\description{
Function to get relative humidity/condense note after cooling
}
\examples{
cooling_air(78, 45, 56, chart_data)
$RH
[1] 97
$note
[1] ""
}

