% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{rhoair_from_tdb_w_pb}
\alias{rhoair_from_tdb_w_pb}
\title{Calculate air density}
\usage{
rhoair_from_tdb_w_pb(tdb, w, pb)
}
\arguments{
\item{tdb}{A numeric vector of dry-bulb temperature in C.}

\item{w}{A numeric vector of humidity ratio in kgWater/kgDryAir.}

\item{pb}{A numeric vector of barometric pressure in Pa.}
}
\value{
A numeric vector of air density (in kg/m3)
}
\description{
Calculate air density
}
\details{
`rhoair_from_tdb_w_pb()` calculates air density from dry-bulb temperature (in
Celsius), humidity ratio (in kgWater/kgDryAir) and barometric pressure (in
Pascals).
}
\examples{
rhoair_from_tdb_w_pb(1, 2, 3)
}
\references{
`PsyRhoAirFnPbTdbW()` in EnergyPlus `Psychrometrics` module.
}
