% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{rhov_from_tdb_rh_lbnd0c}
\alias{rhov_from_tdb_rh_lbnd0c}
\alias{rhov_from_tdb_w_pb}
\alias{rhov_from_tdb_rh}
\title{Calculate vapor density in air}
\usage{
rhov_from_tdb_rh_lbnd0c(tdb, rh)

rhov_from_tdb_w_pb(tdb, w, pb)

rhov_from_tdb_rh(tdb, rh)
}
\arguments{
\item{tdb}{A numeric vector of dry-bulb temperature in C.}

\item{rh}{A numeric vector relative humidity in [0.0, 1.0].}

\item{w}{A numeric vector of humidity ratio in kgWater/kgDryAir.}

\item{pb}{A numeric vector of barometric pressure in Pa.}
}
\value{
A numeric vector of vapor density in air in kg/m3.
}
\description{
Calculate vapor density in air
}
\details{
`rhov_from_tdb_rh_lbnd0c()` calculates the vapor density in air from dry-bulb
temperature (in Celsius) and relative humidity (in 0.0 - 1.0). The name
signifies derivation and temperatures were used with 0C as minimum.

`rhov_from_tdb_w_pb()` calculates vapor density in air from dry-bulb
temperature (in Celsius), humidity ratio (in kgWater/kgDryAir) and barometric
pressure (in Pascals).

`rhov_from_tdb_rh()` calculates vapor density in air from dry-bulb
temperature (in Celcius) and relative humidity (in [0.0 - 1.0]).
}
\examples{
rhov_from_tdb_rh_lbnd0c(13, 0.1)
rhov_from_tdb_w_pb(13, 0.002, 130000)
rhov_from_tdb_rh(25, 0.8)
}
