% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dew_point.R
\name{tdp_from_tdb_twb_pb}
\alias{tdp_from_tdb_twb_pb}
\alias{tdp_from_w_pb}
\title{Calculate dew-point temperature of air}
\usage{
tdp_from_tdb_twb_pb(tdb, twb, pb)

tdp_from_w_pb(w, pb)
}
\arguments{
\item{tdb}{A numeric vector of dry-bulb temperature in C.}

\item{pb}{A numeric vector of barometric pressure in Pa.}
}
\value{
A numeric vector of air dew-point temperature in Celsius.
}
\description{
Calculate dew-point temperature of air
}
\details{
`tdp_from_tdb_twb_pb()` calculates dew-point temperature of air from dry-bulb
temperature (in Celsius), wet-bulb temperature (in Celsius) and barometric
pressure (in Pascals).

`tdp_from_w_pb()` calculates dew-point temperature of air from humidity ratio
(in kgWater/kgDryAir) and barometric pressure (in Pascals).
}
\examples{
tdp_from_tdb_twb_pb(22, 18, 13000)
tdp_from_w_pb(0.002, 13000)
}
