% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enthalpy.R
\name{h_from_tdb_w}
\alias{h_from_tdb_w}
\alias{h_from_tdb_rh_pb}
\title{Calculate enthalpy of air}
\usage{
h_from_tdb_w(tdb, w)

h_from_tdb_rh_pb(tdb, rh, pb)
}
\arguments{
\item{tdb}{A numeric vector of dry-bulb temperature in C.}

\item{w}{A numeric vector of humidity ratio in kgWater/kgDryAir.}

\item{pb}{A numeric vector of barometric pressure in Pa.}
}
\value{
A numeric vector of air enthalpy in J/kg.
}
\description{
Calculate enthalpy of air
}
\details{
`h_from_tdb_w()` calculates enthalpy of air from dry-bulb temperature (in
Celsius) and humidity ratio (in kgWater/kgDryAir).

`h_from_tdb_rh_pb()` calculates enthalpy of air from dry-bulb temperature (in
Celsius), relative humidity (in [0.0, 1.0]) and barometric pressure (in
Pascals).
}
\examples{
h_from_tdb_w(13, 0.002)
h_from_tdb_rh_pb(13, 0.002)
}
