% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Function to pull stored psychrometirc chart data}
\usage{
get_data()
}
\description{
Function to pull stored psychrometirc chart data
}
\examples{
chart_data <- get_data()
}

