% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_capacity.R
\name{cpair_from_w_tdb}
\alias{cpair_from_w_tdb}
\title{Calculate heat capacity of air at constant pressure}
\usage{
cpair_from_w_tdb(tdb, w)
}
\arguments{
\item{tdb}{A numeric vector of dry-bulb temperature in C.}

\item{w}{A numeric vector of humidity ratio in kgWater/kgDryAir.}
}
\value{
A numeric vector of heat capacity of air at constant pressure in
J/(kg.C).
}
\description{
Calculate heat capacity of air at constant pressure
}
\details{
`cpair_from_w_tdb()` calculates the heat capacity of air at constant pressure
from  temperature (in Celsius) and humidity ratio (in kgWater/kgDryAir).
}
\examples{
cpair_from_tdb_w(20, 0.002)
}
