% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatAir.R
\name{HeatAir}
\alias{HeatAir}
\title{Function to get relative humidity after heating}
\usage{
HeatAir(t.pre, rh.pre, t.post, temp.list)
}
\arguments{
\item{t.pre}{Original Dry bulb temperature}

\item{rh.pre}{Original relative humidity}

\item{t.post}{Final Dry bulb temperature}

\item{temp.list}{the stored data from psychrometric chart}
}
\description{
Function to get relative humidity after heating
}
\examples{
HeatAir(45, 40, 85, data.chart)
[1] 10
}

