% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heating_air.R
\name{heating_air}
\alias{heating_air}
\title{Function to get relative humidity after heating}
\usage{
heating_air(T_pre, RH_pre, T_post, temp_list)
}
\arguments{
\item{T_pre}{Original Dry bulb temperature}

\item{RH_pre}{Original relative humidity}

\item{T_post}{Final Dry bulb temperature}

\item{temp_list}{The stored data from psychrometric chart}
}
\description{
Function to get relative humidity after heating
}
\examples{
heating_air(45, 40, 85, chart_data)
[1] 10
}

