% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity_ratio.R
\name{w_from_tdb_h}
\alias{w_from_tdb_h}
\alias{w_from_tdp_pb}
\alias{w_from_tdb_rh_pb}
\alias{w_from_tdb_twb_pb}
\title{Calculate humidity ratio of air}
\usage{
w_from_tdb_h(tdb, h)

w_from_tdp_pb(tdp, pb)

w_from_tdb_rh_pb(tdb, rh, pb)

w_from_tdb_twb_pb(tdb, twb, pb)
}
\arguments{
\item{tdb}{A numeric vector of dry-bulb temperature in C.}

\item{h}{A numeric vector of air enthalpy in J/kg.}

\item{pb}{A numeric vector of barometric pressure in Pa.}
}
\value{
A numeric vector of air humidity ratio in kgWater/kgDryAir.
}
\description{
Calculate humidity ratio of air
}
\details{
`w_from_tdb_h()` calculates humidity ratio of air from dry-bulb temperature
(in Celsius) and enthalpy (in J/kg).

`w_from_tdb_rh_pb()` calculates humidity ratio of air from dry-bulb
temperature (in Celsius), relative humidity (in [0.0, 1.0]) and barometric
pressure (in Pascals).

`w_from_tdb_twb_pb()` calculates humidity ratio of air from dry-bulb
temperature (in Celsius), wet-bulb temperature (in Celsius) and barometric
pressure (in Pascals).

`w_from_tdp_pb()` calculates humidity ratio of air from dew-point temperature
(in Celsius) and barometric pressure (in Pascals).
}
\examples{
w_from_tdb_h(13, 0.002)
w_from_tdp_pb(1, 130000)
w_from_tdb_rh_pb(1, 2, 3)
w_from_tdb_twb_pb(20, 13, 13000)
}
