% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enthalpy.R
\name{hfgair_from_tdb}
\alias{hfgair_from_tdb}
\alias{hgair_from_tdb}
\title{Calculate latent energy of air}
\usage{
hfgair_from_tdb(tdb)

hgair_from_tdb(tdb)
}
\arguments{
\item{tdb}{A numeric vector of dry-bulb temperature in C.}
}
\value{
`hfgair_from_tdb()` returns a numeric vector of heat of vaporization
for moist air in J/kg. `hgair_from_tdb()` returns a numeric vector of latent
energy of the moisture in J/kg.
}
\description{
Calculate latent energy of air
}
\details{
`hfgair_from_tdb()` calculates latent energy of air from temperature (in
Celsius).

`hgair_from_tdb()` calculates latent energy of the moisture as a gas in
the air from temperature (in Celsius).

The enthalpy of a mixture of perfect gases equals the sum of the individual
partial enthalpies of the components. Therefore, the specific enthalpy of
moist air can be calculated as a sum of the specific enthalpy for dry air and
the specific enthalpy for saturated water vapor at the temperature of the
mixture.
}
\examples{
hfgair_from_tdb(13)
hgair_from_tdb(13)
}
