% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_air.R
\name{mix_air}
\alias{mix_air}
\title{Function to get relative humidity and dry_bulb T after mixing}
\usage{
mix_air(T1, RH1, T2, RH2, T1_percent, temp_list)
}
\arguments{
\item{T1}{Dry bulb temperature for first air flow}

\item{RH1}{Relative humidity for first air flow}

\item{T2}{Dry bulb temperature for second air flow}

\item{RH2}{Relative humidity for second air flow}

\item{T1_percent}{Percent of first air flow in the mixed air}

\item{temp_list}{The stored data from psychrometric chart}
}
\description{
Function to get relative humidity and dry_bulb T after mixing
}
\examples{
mix_air(41, 50, 68, 60, 0.375, chart_data)
$Temp
[1] 58
$RH
[1] 64
}

