% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixAir.R
\name{MixAir}
\alias{MixAir}
\title{Function to get relative humidity and dry.bulb t after mixing}
\usage{
MixAir(t1, rh1, t2, rh2, t1.percent, temp.list)
}
\arguments{
\item{t1}{Dry bulb temperature for first air flow}

\item{rh1}{Relative humidity for first air flow}

\item{t2}{Dry bulb temperature for second air flow}

\item{rh2}{Relative humidity for second air flow}

\item{t1.percent}{Percent of first air flow in the mixed air}

\item{temp.list}{the stored data from psychrometric chart}
}
\description{
Function to get relative humidity and dry.bulb t after mixing
}
\examples{
MixAir(41, 50, 68, 60, 0.375, chart.data)
$temp
[1] 58
$rh
[1] 64
}

