% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixMultiAir.R
\name{MixMultiAir}
\alias{MixMultiAir}
\title{Function to get relative humidity and dry.bulb t after multipul mixing}
\usage{
MixMultiAir(vec.t, vec.rh, temp.list)
}
\arguments{
\item{vec.t}{A vector of Dry bulb temperatures}

\item{vec.rh}{A vector of relative humidity corresponding to the vec.t}

\item{temp.list}{the stored data from psychrometric chart}
}
\description{
Function to get relative humidity and dry.bulb t after multipul mixing
}
\examples{
vec.t <- c(40,50,60,70,80)
vec.rh <- c(55,60,75,80,85)
mix.multi.air(vec.t, vec.rh, temp.list)
$temp
[1] 60
$rh
[1] 85
}

