% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_humidity.R
\name{rh_from_tdb_rhov_lbnd0c}
\alias{rh_from_tdb_rhov_lbnd0c}
\alias{rh_from_tdb_rhov}
\alias{rh_from_tdb_w_pb}
\title{Calculate relative humidity of air}
\usage{
rh_from_tdb_rhov_lbnd0c(tdb, rhov)

rh_from_tdb_rhov(tdb, rhov)

rh_from_tdb_w_pb(tdb, w, pb)
}
\arguments{
\item{tdb}{A numeric vector of dry-bulb temperature in C.}

\item{rhov}{A numeric vector of vapor density in kg/m3.}

\item{w}{A numeric vector of humidity ratio in kgWater/kgDryAir.}

\item{pb}{A numeric vector of barometric pressure in Pa.}
}
\value{
A numeric vector of air relative humidity in [0.0 - 1.0].
}
\description{
Calculate relative humidity of air
}
\details{
`rh_from_tdb_rhov_lbnd0c()` calculates relative humidity of air from dry-bulb
temperature (in Celsius) and vapor density (in kg/m3). The name signifies
derivation and temperatures were used with 0C as minimum.

`rh_from_tdb_rhov()` calculates relative humidity of air from dry-bulb
temperature (in Celsius) and vapor density (in kg/m3).

`rh_from_tdb_w_pb()` calculates relative humidity of air from dry-bulb
temperature (in Celsius), humidity ratio (in kgWater/kgDryAir) and barometric
pressure (in Pascals).
}
\examples{
h_from_tdb_w(13, 0.002)
rhov_from_tdb_rhov(25, 0.8)
rhov_from_tdb_rhov(25, 0.8)
}
