% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturation_press.R
\name{psat_from_tdb}
\alias{psat_from_tdb}
\title{Calculate saturation pressure of air}
\usage{
psat_from_tdb(tdb)
}
\arguments{
\item{tdb}{A numeric vector of dry-bulb temperature in C.}
}
\value{
A numeric vector of air saturation pressure in Pascals.
}
\description{
Calculate saturation pressure of air
}
\details{
`psat_from_tdb()` calculates saturation pressure of air from dry-bulb
temperature (in Celsius).
}
\examples{
psat_from_tdb(25)
}
\references{
ASHRAE HANDBOOK OF FUNDAMENTALS, 2005, Chap 6 (Psychrometrics),
Eqn 5 & 6. Compared to Table 3 values (August 2007) with average error of
0.00%, max .30%, min -.39%. (Spreadsheet available on request - Lawrie).
}
