% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturation_temp.R
\name{tsat_from_h_pb}
\alias{tsat_from_h_pb}
\alias{tsat_from_pb}
\title{Calculate saturation pressure of air}
\usage{
tsat_from_h_pb(h, pb)

tsat_from_pb(pb)
}
\arguments{
\item{h}{A numeric vector of air enthalpy in J/(kg.K).}

\item{pb}{A numeric vector of barometric pressure in Pa.}
}
\value{
A numeric vector of air saturation pressure in Pascals.
}
\description{
Calculate saturation pressure of air
}
\details{
`tsat_from_h_pb()` calculates saturation pressure of air from humidity ratio
(in kgWater/kgDryAir) and barometric pressure (in Pascals).

`tsat_from_pb()` calculates saturation temperature of air from barometric
pressure (in Pascals).
}
\examples{
tsat_from_h_pb(13, 13000)
tsat_from_h_pb(13, 13000)
}
