% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_vol.R
\name{v_from_tdb_w_pb}
\alias{v_from_tdb_w_pb}
\title{Calculate specific volume of air}
\usage{
v_from_tdb_w_pb(tdb, w, pb)
}
\arguments{
\item{tdb}{A numeric vector of dry-bulb temperature in C.}

\item{w}{A numeric vector of humidity ratio in kgWater/kgDryAir.}

\item{pb}{A numeric vector of barometric pressure in Pa.}
}
\value{
A numeric vector of air specific volume in m3/kg.
}
\description{
`v_from_tdb_w_pb()` calculates specific volume from dry-bulb temperature (in
Celsius), humidity ratio (in kgWater/kgDryAir) and barometric pressure (in
Pascals).
}
\examples{
v_from_tdb_w_pb(20, 0.02, 13000)
}
