% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wet_bulb.R
\name{twb_from_tdb_w_pb}
\alias{twb_from_tdb_w_pb}
\title{Calculate wet-bulb temperature of air}
\usage{
twb_from_tdb_w_pb(tdb, w, pb)
}
\arguments{
\item{tdb}{A numeric vector of dry-bulb temperature in C.}

\item{w}{A numeric vector of humidity ratio in kgWater/kgDryAir.}

\item{pb}{A numeric vector of barometric pressure in Pa.}
}
\value{
A numeric vector of air wet-bulb temperature in Celsius.
}
\description{
Calculate wet-bulb temperature of air
}
\details{
`h_from_tdb_w()` calculates wet-bulb temperature of air from dry-bulb
temperature (in Celsius), humidity ratio (kgWater/kgDryAir) and barometric
pressure (in Pascals).
}
\examples{
twb_from_tdb_w_pb(13, 0.002, 13000)
}
